Update NN_Stochastic_Descent_Best_Least_Squares Initialised.py with the M to help find the weights W2 and b2 the outer parameters.
In this we have used the library scipy.integrate to import the function quad to integrate between the two piecewise linear functions with the hyper-plane : (w_i).T x = bi for i = 1,..,n

Let phi[0] = ReLU{ (w_0).T x - b0} = 1. 

Which is just meant to be using the Lemma 2.1 from the paper from Cai's. where the it is.

Lemma 2.1:  Assume that the Hyper-planes {(w_i).T x = bi} for i = 1,..,n are distinct . Then { phi[i](x) } for i = 0,1,...,n are linearly independent. 

After using in this code i figured out what the matrix M would be for where it is defined as < 𝜑𝑗(𝐱;𝝎0_𝑗 , 𝑏0_𝑗), 𝜑𝑖(𝐱;𝝎0_𝑖 , 𝑏0_𝑖 > for the 𝐿^2(Ω) inner product. where this is used to help us fine the M_(i,j) matrix element wise where it belongs to ℝ^(𝑛+1)×(𝑛+1). 

Task now:

Now this is found. the only thing to find is the the Projection where Mc = F. Where F is the Projection. 

F_i is defined as < 𝑓= cos(πx) , 𝜑𝑖(𝐱;𝝎0_𝑖 , 𝑏0_𝑖 > for the 𝐿^2(Ω) inner product. Where the we should use a numerical approximation to find integral of a function. Ideally try using the composite trapezoidal rule with the Romberg method for numerical accuracy of the integral.