Update NN_Stochastic_Descent_Best_Least_Squares Initialised.py
Added the projection F using the composite trapezoidal rule with the Romberg integration. 

Then figure out what what c is using the pscedo-inverse as the matrix M came out to be singular. 

Then when we plot the Initual plot it comes out to be very close to the f(x) with a very low cost.

After 1 million iterations in the training we find the NN accurately behaving with a even fare lower cost.